Split structural simplifier.
